feat: convert optinmetrics component to tsx
